// type-generator/generated-client/models/FileUploadResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FileUploadResponseDto = {
    /**
     * Upload success status
     */
    success: boolean;
    /**
     * File URL
     */
    url: string;
    /**
     * Original filename
     */
    originalName: string;
    /**
     * Stored filename
     */
    filename: string;
    /**
     * File size in bytes
     */
    size: number;
    /**
     * MIME type
     */
    mimetype: string;
    /**
     * Upload message
     */
    message: string;
};